Fixing bug with HAML blocks and Rails 3 - Resolves Issue #6
with_output_buffer is actually not a public method in Rails, so I probably shouldn't have used it anyway. capture, on the other hand, is public, and (currently) utilizes with_output_buffer in its implemenation.

As it turns out, Rails 3 also deprecates block helpers that don't have the "=" sign. Since the table.data method accepts an optional block parameter, and seems to be outputting content, it really should called with an "=" sign, in both erb and haml, and not with a '-' in haml. Also this form would (and should) be valid:

  <%= table_for(@posts) do |table|
    table.data(:actions => :all)
  end %>